Make test folder for src structure and refactored verify_version tests
